Improvements in HPE Contracts
* Filtered Companies on Contract Creation Screen
* Added Logos and Flags for Companies & Countries
* Added User Profile Based Companies and Templates
* Added Contract Completeness in Contracts Listing
* Added Contract Template Policy
* Added Smart Behavior of Grouped Rows
* Added Contract Copying
* Added User's profile country for contract creation
* Added Role for Internal Companies to be then used in the profile
* Added HPE Contract Unit Test
* Fixed issues with Vendor relationships with controller
* Fixed missing declaration of profile Helper
* Prevented Group description rows in schedule payments
* Updated PDF Parser Service to handle line description gaps

+++ app/Services/HpeContractFileService.php
@@ -42,9 +42,9 @@ class HpeContractFileService
     public function processImport(HpeContractFile $hpeContractFile): ImportResponse
     {
         try {
-            // if (null !== $hpeContractFile->imported_at) {
-            //     return new ImportResponse(true);
-            // }
+            if (null !== $hpeContractFile->imported_at) {
+                return new ImportResponse(true);
+            }
 
             return DB::transaction(function () use ($hpeContractFile) {
                 $hpeContractFile->hpeContractData()->delete();
